move metricsStorage to internals
